I started working on the onboarding screen. Still need to add some more features and tweak the layout a little bit.
I believe all details have been included, but
I still need to test the code tomorrow to see how
it actually looks. Also we still need functionality
to be added to some of the buttons as well as the
back end.

For now I just copied the layout of the
given onboarding screen.

Slide 5 of Project Introduction

- [] I have tested my code on a different machine and confirmed that it works as expected.