Beginning Game/Player Action Screen
Began building all of the components for the player action screen.  Using a grid based approach with 12 columns and 16 rows currently.  I havent been able to nail the formatting yet.  Will finish up tomorrow and begin querying back end for values.

This approach is to componentize as much as possible, so the playerDisplay is one component, the playerAction is one component, and will need to create a countdown component as well.